Split testimonios out of the IPC monolith
Same three-block move as prosopography, one domain wider:

  electron/ipc.ts     5239-5450 -> electron/ipc/testimonies.ts     (74 handlers)
  electron/preload.ts 1537-1628 -> electron/preload/testimonies.ts (74 bindings)
  shared/types.ts     9040-9130 -> shared/api/testimonies.ts       (74 signatures)

shared/types.ts declares 41 of the Testimony* types itself rather than importing
them, so the slice imports them back from ../types. That cycle is types-only and
erased at build time.

electron/ipc.ts is down from 6,013 lines to 5,578, electron/preload.ts from 1,862
to 1,725, shared/types.ts from 9,553 to 9,356.

Verified: 1,425 tests pass, the IPC contract test confirms all 74 channels are
still registered exactly once and reachable, and npm run test:e2e (the suite CI
gates on) passes. npm run test:e2e:testimony fails, but it fails identically on
the pristine branch point — the guided-tour overlay from tourEngine.tsx covers the
sidebar and swallows the click, because the script suppresses what's-new modals by
a hard-coded key list that does not cover the per-vault tour. Unrelated to IPC,
and not gated by CI.
